tests/requirements.txt
py==1.4.32
pytest==3.0.5
python-bitcoinrpc==1.0
